requirements.txt
python-dotenv==1.0.0
py-cord==2.4.1
docker==6.1.2